feat(szamlazz-agent, szamlazz-adatkapcsolat)!: one integer-width policy (`i64`) for the XSD ints; `InvoiceAppearance` one width and one JSON shape; the receiver's `tipus` is `document_type`; ADR 0010 keeps the two models (#129)
Every integer element of a queried or pushed document is an `i64` in both
crates (the agent's `alap/id`, `gazdEsemAzon`, `forras`, the parties'
`id`, `lokacio`, `sztetordering`, `afalevon`, `banktranzid`, the
`szlahu_id` header, a receipt's `id`; the receiver's `alap/id` and
`lokacio`, and the `InvoiceAck` id), signed and wide whatever the schema
declares; each crate's docs carry the rule (*Integer widths*).
`InvoiceAppearance` is `i64` in both and serialises as the integer code
in both (the agent's string `"2"` is gone; the CLI's JSON changes with
it); the receiver's `is_e_invoice` is the `Electronic` variant. The
receiver's `InvoiceInfo::kind` / `e_invoice`, `BankTransaction::kind` and
`ReceiptInfo::kind` are `document_type` / `appearance` /
`transaction_type` / `document_type`, the archived JSON included.

ADR 0010 records the shared-model decision: two models kept
deliberately (strict vs lenient parse are two products of one schema;
the receiver takes no agent dependency), with the width policy, the
open-set rule and one JSON shape as the shared vocabulary. The worker's
`FoundDocument` / `IssuedDocument` follow the width (ADR 0009 allows the
reshape). Both READMEs list the breaking changes under 0.4.

Closes #129.

+++ crates/restate-szamlazz/src/gateway/document.rs
@@ -44,8 +44,9 @@ pub struct FoundDocument {
     /// identifier, not an account's or a seller's. Read by no handler;
     /// carried so that a journal entry names the document the way
     /// szamlazz.hu's own records do (the same value a create reply's
-    /// `szlahu_id` header carries, [`IssuedDocument::document_id`]).
-    pub document_id: u64,
+    /// `szlahu_id` header carries, [`IssuedDocument::document_id`]). An
+    /// `i64`, the agent crate's width for every XSD integer.
+    pub document_id: i64,
     /// The document number (`szamlaszam`).
     pub number: String,
     /// The document type (`tipus`): `SZ` invoice, `D` proforma, `ES`
@@ -74,8 +75,9 @@ pub struct FoundDocument {
     /// The `eszamla` code as szamlazz.hu reports it: `0` not an invoice (a
     /// proforma), `1` paper, `2`/`3` e-invoice; the agent crate's
     /// [`InvoiceAppearance`] reads it, and [`Self::e_invoice`] is the reading
-    /// the storno handlers take.
-    pub appearance: i32,
+    /// the storno handlers take. An `i64`, the code's width in the agent
+    /// crate.
+    pub appearance: i64,
     /// The issue date (`kelt`).
     pub issue_date: Option<Date>,
     /// The fulfillment date (`telj`): what a storno of this document must
@@ -284,7 +286,7 @@ pub struct IssuedDocument {
     /// the same value a query returns as [`FoundDocument::document_id`];
     /// `None` when the header is absent or not a number. Read by no handler;
     /// carried for the same correlation.
-    pub document_id: Option<u64>,
+    pub document_id: Option<i64>,
     /// The net total (`szamlanetto`).
     pub net_total: Option<Decimal>,
     /// The gross total (`szamlabrutto`).
